🚚 Rename : "패키지명 climatecocktail로 변경"
Ref : #96
