+++ src/main/java/alcoholboot/toastit/feature/climatecocktail/dto/WeatherItemDTO.java
@@ -1,7 +1,5 @@
-package alcoholboot.toastit.feature.api.dto;
+package alcoholboot.toastit.feature.climatecocktail.dto;
 
-import alcoholboot.toastit.feature.api.entity.WeatherEntity;
-import lombok.Builder;
 import lombok.Data;
 
 @Data
