+++ src/main/java/alcoholboot/toastit/feature/climatecocktail/dto/WeatherResponseDTO.java
@@ -1,4 +1,4 @@
-package alcoholboot.toastit.feature.api.dto;
+package alcoholboot.toastit.feature.climatecocktail.dto;
 
 import lombok.Data;
