+++ src/main/java/alcoholboot/toastit/feature/climatecocktail/entity/WeatherEntity.java
@@ -1,4 +1,4 @@
-package alcoholboot.toastit.feature.api.entity;
+package alcoholboot.toastit.feature.climatecocktail.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
